modify BlockFee struct

+++ model/fee.go
@@ -1,7 +1,7 @@
 package model
 
 type BlockFee struct {
-	FastestFee	uint64	`json:"fastestFee"`
-	HalfHourFee	uint64	`json:"halfHourFee"`
-	HourFee		uint64	`json:"hourFee"`
+	FastestFee	int	`json:"fastestFee"`
+	HalfHourFee	int	`json:"halfHourFee"`
+	HourFee		int	`json:"hourFee"`
 }
